requirements.txt
# Polymarket Analyzer - Dependencies
# Production-quality prediction market analysis toolkit

# =============================================================================
# Core Platform SDKs
# =============================================================================
py-clob-client>=0.34.1          # Polymarket official CLOB client
kalshi-python>=2.1.4            # Kalshi official SDK
polymarket-apis>=0.1.0          # Unified Polymarket APIs (CLOB, Gamma, GraphQL)

# =============================================================================
# Data & Analysis
# =============================================================================
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0                   # Statistical analysis, distributions

# =============================================================================
# Async & HTTP
# =============================================================================
httpx>=0.24.0                   # Modern async HTTP client
websockets>=11.0                # WebSocket support
aiohttp>=3.8.0                  # Async HTTP for streaming

# =============================================================================
# Infrastructure
# =============================================================================
structlog>=23.1.0               # Structured logging
pydantic>=2.0.0                 # Data validation and settings
python-dotenv>=1.0.0            # Environment variable management
click>=8.1.0                    # CLI framework
aiosqlite>=0.19.0               # Async SQLite for data persistence

# =============================================================================
# Blockchain/Web3 (for Polymarket trading)
# =============================================================================
web3>=6.0.0                     # Ethereum/Polygon interaction
eth-account>=0.9.0              # Account management for signing

# =============================================================================
# Visualization (for analysis)
# =============================================================================
matplotlib>=3.7.0
plotly>=5.15.0

# =============================================================================
# Notebooks
# =============================================================================
jupyter>=1.0.0
nest-asyncio>=1.5.0                # Async support in notebooks

# =============================================================================
# Testing
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
